fix .dtype in preproces_image

+++ app/app_main.py
@@ -44,8 +44,10 @@ def preprocess_image(image_bytes):
 
         img_data = np.array(img).transpose(2, 0, 1).astype('float32') / 255.0
 
-        mean = np.array([0.485, 0.456, 0.406]).reshape(3, 1, 1)
-        std = np.array([0.229, 0.224, 0.225]).reshape(3, 1, 1)
+        mean = np.array([0.485, 0.456, 0.406],
+                        dtype=np.float32).reshape(3, 1, 1)
+        std = np.array([0.229, 0.224, 0.225],
+                       dtype=np.float32).reshape(3, 1, 1)
         img_data = (img_data - mean) / std
 
         return np.expand_dims(img_data, axis=0)
